move layers component over to hooks
