+++ src/components/Layers/constants.ts
@@ -1,0 +1,6 @@
+// Node type constants for the tree structure
+export const NODE_TRACKS = 'node-tracks'
+export const NODE_FIELDS = 'node-fields'
+export const NODE_ZONES = 'node-zones'
+export const NODE_POINTS = 'node-points'
+export const NODE_BACKDROPS = 'node-backdrops'
